[#4] refactor: Logo 컴포넌트 common으로 분류
