+++ src/components/common/Logo/style.ts
