+++ src/components/common/Navbar/index.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react';
 import { Link } from 'react-router-dom';
 
-import Logo from '@/components/Logo';
+import Logo from '../Logo';
 import Uri from '@/constants/uri';
 
 import { navStyle, yellowLineButton, yellowFullButton } from './style';
